Centralized hasProtoExtra matcher
