+++ app/BUILD.bazel
@@ -832,6 +832,7 @@ kt_android_library(
         "src/sharedTest/java/org/oppia/android/app/utility/OrientationChangeAction.kt",
         "src/sharedTest/java/org/oppia/android/app/utility/ProgressMatcher.kt",
         "src/sharedTest/java/org/oppia/android/app/utility/TabMatcher.kt",
+        "src/sharedTest/java/org/oppia/android/app/utility/ProtoExtraMatcher.kt",
     ],
     visibility = [
         ":app_testing_visibility",
