+++ scripts/assets/kdoc_validity_exemptions.textproto
@@ -271,6 +271,7 @@ exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/DragV
 exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/DrawableMatcher.kt"
 exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/EspressoTestsMatchers.kt"
 exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/OrientationChangeAction.kt"
+exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/ProtoExtraMatcher.kt"
 exempted_file_path: "app/src/sharedTest/java/org/oppia/android/app/utility/TabMatcher.kt"
 exempted_file_path: "data/src/main/java/org/oppia/android/data/backends/gae/api/FeedbackReportingService.kt"
 exempted_file_path: "domain/src/main/java/org/oppia/android/domain/audio/AudioPlayerController.kt"
